PIMS-1427 Adding property endpoints (#152)
Refresh DB Migration - adding properties view
Extending unit tests
Fixing paged endpoints

+++ backend/entities/BuildingEvaluation.cs
@@ -37,7 +37,7 @@ namespace Pims.Dal.Entities
         /// get/set - The value of the fiscal key for this building.
         /// </summary>
         /// <value></value>
-        public float Value { get; set; }
+        public decimal Value { get; set; }
 
         /// <summary>
         /// get/set - A note related to this fiscal value.
@@ -59,7 +59,7 @@ namespace Pims.Dal.Entities
         /// <param name="date"></param>
         /// <param name="key"></param>
         /// <param name="value"></param>
-        public BuildingEvaluation(Building building, DateTime date, EvaluationKeys key, float value)
+        public BuildingEvaluation(Building building, DateTime date, EvaluationKeys key, decimal value)
         {
             this.BuildingId = building?.Id ??
                 throw new ArgumentNullException(nameof(building));
